d_a_npc_zra mostly done, misc npc work (#2140)

+++ include/d/meter/d_meter2_info.h
@@ -154,8 +154,8 @@ public:
     void setMsgTimeMs(u32 msgTime) { mMsgTimeMs = msgTime; }
     void setTimeMs(s32 i_time) { mTimeMs = i_time; }
     u32 getTimeMs() { return mTimeMs; }
-    u8 getNowCount() { return mNowCount; }
-    u8 getMaxCount() { return mMaxCount; }
+    int getNowCount() { return mNowCount; }
+    int getMaxCount() { return mMaxCount; }
     void setScopeZoomPointer(u8 param_0) { mScopeZoomPointer = param_0; }
     u8 getItemExplainWindowStatus() { return mItemExplainWindowStatus; }
     void setItemExplainWindowStatus(u8 i_value) { mItemExplainWindowStatus = i_value; }
@@ -611,11 +611,11 @@ inline void dMeter2Info_setTimeMs(s32 i_time) {
     g_meter2_info.setTimeMs(i_time);
 }
 
-inline u8 dMeter2Info_getNowCount() {
+inline int dMeter2Info_getNowCount() {
     return g_meter2_info.getNowCount();
 }
 
-inline u8 dMeter2Info_getMaxCount() {
+inline int dMeter2Info_getMaxCount() {
     return g_meter2_info.getMaxCount();
 }
